Relocate the methods of the App class
