+++ person.rb
@@ -2,6 +2,8 @@ require './nameable'
 require './capitalize_decorator'
 require './trimmer_decorator'
 require './rental'
+require './student'
+require './teacher'
 
 class Person < Nameable
   # Getters and setters
@@ -27,6 +29,32 @@ class Person < Nameable
 
   def add_rental(book, date)
     Rental.new(date, self, book)
+  end
+
+  # List all people.
+  def self.list_people(people)
+    if people.length.zero?
+      puts 'There are not people at the moment.'
+    else
+      people.each_with_index do |person, idx|
+        puts "#{idx}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+      end
+    end
+  end
+
+  # Create a person (teacher or student, not a plain Person).
+  def create_person
+    person_type = 0
+    until [1, 2].include?(person_type)
+      print 'Do you want to create a student (1) or do you want to create a teacher (2)? [Input the number]:'
+      person_type = gets.chomp.to_i
+    end
+    case person_type
+    when 1 then Student.create_student
+    when 2 then Teacher.create_teacher
+    end
+    person = person_type == 1 ? 'Student' : 'Teacher'
+    print "#{person} created successfully\n"
   end
 
   # Private methods
