+++ student.rb
@@ -15,4 +15,17 @@ class Student < Person
     @classroom = classroom
     @classroom.push(self) unless @classroom.students.include?(self)
   end
+
+  def self.create_student
+    print 'Age: '
+    age = gets.chomp.to_i
+    print 'Name: '
+    name = gets.chomp
+    permission = ''
+    until %w[Y N].include?(permission)
+      print 'Has parrent permission? [Y/N]:'
+      permission = gets.chomp.capitalize
+    end
+    Student.new(nil, age, name, parent_permission: permission == 'Y')
+  end
 end
