+++ teacher.rb
@@ -9,4 +9,14 @@ class Teacher < Person
   def can_use_services?
     true
   end
+
+  def self.create_teacher
+    print 'Age: '
+    age = gets.chomp.to_i
+    print 'Name: '
+    name = gets.chomp
+    print 'Specialization: '
+    specialization = gets.chomp.capitalize
+    Teacher.new(specialization, age, name)
+  end
 end
